Weather Project (junior year)
Created a weather app using real-time data from a weather API. When you enter the zip code, you get the month, day, time (in EST), the actual temperature, the psuedo temperature (how it feels), and the weather. There is a discrete seek bar where you could see the temperatures throughout that day. The theme of the project is a hacker/matrix based theme.