Hackathon Project (2023-2024)
For the 2023-2024 HACKSB, my team and I created an app called Revive, which is dedicated to helping with mental health, substance abuse, and physical therapy. I also created a chatbot dedicated to it to learn more about mental health, substance abuse, and physical therapy. We happily won first place at the hackathon due to the creativity of the project.